+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/extensions/ChannelExtension.kt
@@ -23,6 +23,7 @@ import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.models.Channel
 import io.getstream.chat.android.models.Member
 import io.getstream.chat.android.models.User
+import io.getstream.chat.android.models.UserId
 
 public fun Channel.isAnonymousChannel(): Boolean = id.isAnonymousChannelId()
 
@@ -97,10 +98,19 @@ public fun Channel.countUnreadMentionsForUser(user: User): Int {
  *
  * @return The number of unread messages in the channel for the current user.
  */
-public val Channel.currentUserUnreadCount: Int
-    get() = ChatClient.instance().getCurrentUser()?.let { currentUser ->
-        read.firstOrNull { it.user.id == currentUser.id }?.unreadMessages
-    } ?: 0
+public fun Channel.currentUserUnreadCount(
+    currentUserId: UserId? = ChatClient.instance().getCurrentUser()?.id,
+): Int = read.firstOrNull { it.user.id == currentUserId }?.unreadMessages ?: 0
 
-public fun Channel.syncUnreadCountWithReads(): Channel =
-    copy(unreadCount = currentUserUnreadCount)
+/**
+ * Synchronizes the unread count of the channel with the read state of the current user.
+ *
+ * @param currentUserId The ID of the current user. If not provided, the ID of the current user will be taken
+ * from the ChatClient if it is available.
+ *
+ * @return A new instance of the channel with the unread count synchronized with the read state of the current user.
+ */
+public fun Channel.syncUnreadCountWithReads(
+    currentUserId: UserId? = ChatClient.instance().getCurrentUser()?.id,
+): Channel =
+    copy(unreadCount = currentUserUnreadCount(currentUserId))