+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/DownstreamChannelDtoAdapterTest.kt
@@ -24,7 +24,7 @@ import org.amshove.kluent.shouldThrow
 import org.junit.jupiter.api.Test
 
 internal class DownstreamChannelDtoAdapterTest {
-    private val parser = MoshiChatParser { "" }
+    private val parser = ParserFactory.createMoshiChatParser()
 
     @Test
     fun `Deserialize JSON channel with custom fields`() {