+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/UpstreamMessageDtoAdapterTest.kt
@@ -27,7 +27,7 @@ import org.amshove.kluent.shouldThrow
 import org.junit.jupiter.api.Test
 
 internal class UpstreamMessageDtoAdapterTest {
-    private val parser = MoshiChatParser { "" }
+    private val parser = ParserFactory.createMoshiChatParser()
 
     @Test
     fun `Serialize JSON message with custom fields`() {