+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/UpstreamReactionDtoAdapterTest.kt
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Test
 
 internal class UpstreamReactionDtoAdapterTest {
 
-    private val parser = MoshiChatParser { "" }
+    private val parser = ParserFactory.createMoshiChatParser()
 
     @Test
     fun `Serialize JSON reaction with custom fields`() {