+++ stream-chat-android-docs/src/main/java/io/getstream/chat/docs/java/client/cms/UnreadCounts.java
@@ -104,7 +104,10 @@ public class UnreadCounts {
             client.queryChannel("channel-type", "channel-id", queryChannelRequest, false).enqueue((result) -> {
                 if (result.isSuccess()) {
                     // Unread count for the current user
-                    Integer unreadCount = ChannelExtensionKt.getCurrentUserUnreadCount(result.getOrNull());
+                    Integer unreadCount = ChannelExtensionKt.currentUserUnreadCount(
+                            result.getOrNull(),
+                            client.getCurrentUser().getId()
+                    );
                 } else {
                     // Handle error
                 }